refactor: help/http + help/source (mirrors legado)
Match legado Android help/ structure:
  core/network/HttpClient.ts    → help/http/HttpClient.ts
  core/network/CookieStore.ts   → help/http/CookieStore.ts
  core/network/BookSourceImporter.ts → help/source/BookSourceImporter.ts
  core/network/RssSourceImporter.ts  → help/source/RssSourceImporter.ts

Pure rename — no logic changes.
